Added Usuaria class as entity
This commit introduces the 'Usuaria' class as an entity, mapping it to the PostgreSQL database. Annotations such as @Entity, @Table, and @Id were used to define its persistence behavior. Necessary configurations for the database connection were set in the application.properties file. This marks the beginning of user management functionality within the project.